Moved plugin to different package.
