+++ stone-java-gradle-plugin/src/main/kotlin/com/dropbox/stone/java/model/ClientSpec.kt
@@ -1,4 +1,4 @@
-package plugins.model
+package com.dropbox.stone.java.model
 
 import java.io.Serializable
